refactor(ingress): host TableName/ColumnName in the buffer module
The QWP merge (#141) moved these public types into the private `ilp`
submodule and re-exported them upward via `pub use self::ilp::...`.
That introduced a rustdoc intra-doc ambiguity: from inside `ilp.rs`,
the `[Buffer]` link in each type's own doc resolved to the internal
`ilp::Buffer` (the ILP-specific implementation buffer) rather than the
user-facing `ingress::Buffer`. Move both types — plus their `impl`,
`TryFrom`, `AsRef` blocks, and validator tests — back into `buffer.rs`,
where they were before the QWP refactor and where they are still
publicly exported from. Drop the now-redundant `pub use self::ilp::...`
re-export. Public path `questdb::ingress::{TableName, ColumnName}` is
unchanged. Also fixes a pre-existing doc typo on `ColumnName::new`
("Construct a validated table name" -> "column name").
